data: Move file review line validation
Pathless file-review line actions parse the requested IDs, match them
against shown review selections, and raise stale-review guidance.

That flow lived inside action scope, alongside command scope resolution and
whole-file refusal handling. The mix forced action scope to depend on line
parsing and review-selection validation.

This commit addresses that by moving pathless line-action validation and
its stale-review exception into a dedicated file-review validation module.

Subsequent test commits will update direct validation imports and add
boundary coverage.
