+++ src/git_stage_batch/data/file_review/action_scope.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import shlex
 
-from ...core.line_selection import parse_line_selection_ranges
 from ...exceptions import CommandError
 from ...i18n import _
 from . import records as _records
@@ -19,9 +18,9 @@ from .freshness import (
     selected_change_kind_matches_review_source as _selected_change_kind_matches_review_source,
     selected_change_matches_review_state as _selected_change_matches_review_state,
 )
-from .selection_validation import (
-    shown_review_selections_for_action as _shown_review_selections_for_action,
-    validate_review_scoped_line_selection as _validate_review_scoped_line_selection,
+from .line_action_validation import (
+    raise_stale_or_mismatched_file_review as _raise_stale_or_mismatched_file_review,
+    validate_pathless_review_line_action as _validate_pathless_review_line_action,
 )
 from .state import (
     clear_last_file_review_state,
@@ -29,10 +28,6 @@ from .state import (
     read_last_file_review_state,
 )
 from ..selected_change.store import SelectedChangeKind, read_selected_change_kind
-
-
-class ReviewScopedSelectionError(CommandError):
-    """Raised when a pathless line action is not valid for the current review."""
 
 
 def _get_selected_change_file_path() -> str | None:
@@ -104,7 +99,7 @@ def resolve_batch_source_action_scope(
             )
             return _records.ActionScopeResolution(file=reviewed_file if reviewed_file is not None else file)
 
-        review_state = validate_pathless_review_line_action(
+        review_state = _validate_pathless_review_line_action(
             review_action,
             line_ids,
             source=_records.ReviewSource.BATCH,
@@ -139,7 +134,7 @@ def resolve_batch_source_action_scope(
             )
             return _records.ActionScopeResolution(file=reviewed_file if reviewed_file is not None else file)
 
-        review_state = validate_pathless_review_line_action(
+        review_state = _validate_pathless_review_line_action(
             review_action,
             line_ids,
             source=_records.ReviewSource.BATCH,
@@ -159,18 +154,6 @@ def _format_pages(pages: set[int]) -> str:
     return format_line_ids(sorted(pages))
 
 
-def _print_stale_or_mismatched_file_review_help(action: str, review_state: _records.FileReviewState) -> None:
-    show_command = _show_command_for_review_state(review_state)
-    raise ReviewScopedSelectionError(
-        _(
-            "The file review for {file} no longer matches the selected file view.\n"
-            "Line IDs may no longer match.\n\n"
-            "Run:\n"
-            "  {command}"
-        ).format(file=review_state.file_path, command=show_command)
-    )
-
-
 def refuse_live_action_for_batch_selection(action: _records.FileReviewAction | str) -> bool:
     """Refuse bare live actions when the current selection came from a batch view."""
     review_action = _records.coerce_review_action(action)
@@ -184,9 +167,9 @@ def refuse_live_action_for_batch_selection(action: _records.FileReviewAction | s
     review_state = read_last_file_review_state()
     if review_state is not None:
         if review_state.source != _records.ReviewSource.BATCH:
-            _print_stale_or_mismatched_file_review_help(review_action.value, review_state)
+            _raise_stale_or_mismatched_file_review(review_state)
         if not _selected_change_matches_review_state(review_state):
-            _print_stale_or_mismatched_file_review_help(review_action.value, review_state)
+            _raise_stale_or_mismatched_file_review(review_state)
 
         lines = [
             _("The selected file view for {file} came from batch '{batch}', not the live working tree.").format(
@@ -249,12 +232,12 @@ def refuse_ambiguous_bare_action_after_partial_file_review(action: _records.File
             SelectedChangeKind.BATCH_BINARY,
             SelectedChangeKind.BATCH_GITLINK,
         ):
-            _print_stale_or_mismatched_file_review_help(review_action.value, review_state)
+            _raise_stale_or_mismatched_file_review(review_state)
         clear_last_file_review_state()
         return False
 
     if not _review_state_matches_action(review_state, review_action):
-        _print_stale_or_mismatched_file_review_help(review_action.value, review_state)
+        _raise_stale_or_mismatched_file_review(review_state)
 
     if review_state.entire_file_shown:
         return False
@@ -364,7 +347,7 @@ def validate_implicit_live_to_batch_file_action(
         )
     if refuse_live_action_for_batch_selection(review_action):
         return _records.ImplicitLiveToBatchFileActionResult(should_stop=True)
-    review_state = validate_pathless_review_line_action(
+    review_state = _validate_pathless_review_line_action(
         review_action,
         line_id_specification,
         source=_records.ReviewSource.FILE_VS_HEAD,
@@ -404,7 +387,7 @@ def resolve_live_to_batch_action_scope(
                 source=_records.ReviewSource.FILE_VS_HEAD,
             )
             return _records.ActionScopeResolution(file=reviewed_file if reviewed_file is not None else file)
-        review_state = validate_pathless_review_line_action(
+        review_state = _validate_pathless_review_line_action(
             review_action,
             line_ids,
             source=_records.ReviewSource.FILE_VS_HEAD,
@@ -457,7 +440,7 @@ def resolve_live_line_action_scope(
     refuse_bare_action_after_auto_advance_disabled(action_command)
 
     if file is None and validate_pathless_before_live_guard:
-        review_state = validate_pathless_review_line_action(
+        review_state = _validate_pathless_review_line_action(
             review_action,
             line_id_specification,
             source=source,
@@ -470,72 +453,10 @@ def resolve_live_line_action_scope(
     if refuse_live_action_for_batch_selection(review_action):
         return _records.ActionScopeResolution(file=file, should_stop=True)
 
-    review_state = validate_pathless_review_line_action(
+    review_state = _validate_pathless_review_line_action(
         review_action,
         line_id_specification,
         source=source,
         batch_name=batch_name,
     )
     return _records.ActionScopeResolution(file=file, review_state=review_state)
-
-
-def validate_pathless_review_line_action(
-    action: _records.FileReviewAction | str,
-    line_id_specification: str,
-    *,
-    source: _records.ReviewSource | str | None = None,
-    batch_name: str | None = None,
-) -> _records.FileReviewState | None:
-    """Validate pathless --line against the last file review."""
-    review_action = _records.coerce_review_action(action)
-    review_state = read_last_file_review_state()
-    if review_state is None:
-        return None
-    if (
-        source is not None
-        and review_state.source != _records.coerce_review_source(source)
-    ):
-        _print_stale_or_mismatched_file_review_help(review_action.value, review_state)
-    if batch_name is not None and review_state.batch_name != batch_name:
-        _print_stale_or_mismatched_file_review_help(review_action.value, review_state)
-    if not _review_state_matches_action(review_state, review_action):
-        _print_stale_or_mismatched_file_review_help(review_action.value, review_state)
-    if (
-        review_state.source == _records.ReviewSource.BATCH
-        and review_action in (
-            _records.FileReviewAction.INCLUDE,
-            _records.FileReviewAction.SKIP,
-            _records.FileReviewAction.DISCARD,
-            _records.FileReviewAction.INCLUDE_TO_BATCH,
-            _records.FileReviewAction.DISCARD_TO_BATCH,
-        )
-        and not any(review_action in selection.actions for selection in review_state.selections)
-    ):
-        lines = [
-            _("The selected file view for {file} came from batch '{batch}', not the live working tree.").format(
-                file=review_state.file_path,
-                batch=review_state.batch_name,
-            )
-        ]
-        line_command = _line_action_command(review_action, review_state, line_spec=line_id_specification)
-        if line_command is not None:
-            lines.extend(["", _("To act on the batch file:"), f"  {line_command}"])
-        else:
-            lines.extend(
-                [
-                    "",
-                    _("Batch reviews do not support this action."),
-                    _("If you meant to act on live working-tree changes, open a live file review:"),
-                    f"  git-stage-batch show --file {shlex.quote(review_state.file_path)}",
-                ]
-            )
-        raise CommandError("\n".join(lines))
-
-    try:
-        requested_ids = parse_line_selection_ranges(line_id_specification)
-    except ValueError as error:
-        raise CommandError(str(error)) from error
-
-    valid_selections = _shown_review_selections_for_action(review_state, review_action)
-    _validate_review_scoped_line_selection(requested_ids, valid_selections)
-    return review_state
